Atualização de revendas e admin

+++ database/migrations/2018_07_23_233529_create_planos_table.php
@@ -18,7 +18,7 @@ class CreatePlanosTable extends Migration
             $table->string('nome');
             $table->longText('descricao');
             $table->integer('anuncios');
-            $table->double('preco');
+            $table->decimal('preco');
             $table->timestamps();
         });
     }
